splited bf and batch itertor declaration and impl, and added a speacial include file for brute force alg, to avoid circular includes
