+++ src/VecSim/algorithms/brute_force/brute_force_factory.cpp
@@ -1,8 +1,6 @@
 #include "VecSim/algorithms/brute_force/brute_force_factory.h"
-#include "VecSim/algorithms/brute_force/brute_force.h"
 #include "VecSim/algorithms/brute_force/brute_force_single.h"
 #include "VecSim/vec_sim_common.h" // labelType
-#include "VecSim/algorithms/brute_force/bf_batch_iterator.h"
 
 VecSimIndex *BruteForceFactory::NewIndex(const BFParams *params,
                                          std::shared_ptr<VecSimAllocator> allocator) {
@@ -29,14 +27,4 @@ size_t BruteForceFactory::EstimateInitialSize(const BFParams *params) {
 
 size_t BruteForceFactory::EstimateElementSize(const BFParams *params) {
     return params->dim * VecSimType_sizeof(params->type) + sizeof(labelType);
-}
-
-// TODO overload for doubles
-VecSimBatchIterator *
-BruteForceFactory::newBatchIterator(void *queryBlob, VecSimQueryParams *queryParams,
-                                    std::shared_ptr<VecSimAllocator> allocator,
-                                    const BruteForceIndex<float, float> *index) {
-
-    // Ownership of queryBlobCopy moves to BF_BatchIterator that will free it at the end.
-    return new (allocator) BF_BatchIterator<float, float>(queryBlob, index, queryParams, allocator);
 }
